renamed darcyEq.py -> darcy.py
